backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
kombu==5.3.4
pymongo==4.6.1
python-dotenv==1.0.0
pydantic==2.5.2
eventlet==0.33.3
slowapi==0.1.8
prometheus_client==0.19.0
requests==2.31.0
pandas==2.2.0
numpy==1.26.3
python-dateutil==2.8.2 